move to test cotainers and change package name
