+++ server/src/main/java/com/objectcomputing/checkins/services/member_skill/MemberSkillAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.objectcomputing.checkins.services.memberSkill;
+package com.objectcomputing.checkins.services.member_skill;
 
 public class MemberSkillAlreadyExistsException extends RuntimeException {
     public MemberSkillAlreadyExistsException(String message) {
