+++ server/src/test/java/com/objectcomputing/checkins/services/member_skill/MemberSkillControllerTest.java
@@ -1,0 +1,275 @@
+package com.objectcomputing.checkins.services.member_skill;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import com.objectcomputing.checkins.services.TestContainersSuite;
+import com.objectcomputing.checkins.services.fixture.MemberProfileFixture;
+import com.objectcomputing.checkins.services.fixture.MemberSkillFixture;
+import com.objectcomputing.checkins.services.fixture.SkillFixture;
+import com.objectcomputing.checkins.services.memberprofile.MemberProfile;
+import com.objectcomputing.checkins.services.skills.Skill;
+import io.micronaut.core.type.Argument;
+import io.micronaut.http.HttpRequest;
+import io.micronaut.http.HttpResponse;
+import io.micronaut.http.HttpStatus;
+import io.micronaut.http.client.HttpClient;
+import io.micronaut.http.client.annotation.Client;
+import io.micronaut.http.client.exceptions.HttpClientResponseException;
+import org.junit.jupiter.api.Test;
+
+import javax.inject.Inject;
+import java.util.*;
+import java.util.stream.Collectors;
+
+import static com.objectcomputing.checkins.services.role.RoleType.Constants.ADMIN_ROLE;
+import static com.objectcomputing.checkins.services.role.RoleType.Constants.MEMBER_ROLE;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+public class MemberSkillControllerTest extends TestContainersSuite implements MemberProfileFixture, SkillFixture, MemberSkillFixture {
+
+    @Inject
+    @Client("/services/member-skill")
+    HttpClient client;
+
+
+    @Test
+    void testCreateAMemberSkill() {
+
+        MemberProfile memberProfile = createADefaultMemberProfile();
+        Skill skill = createADefaultSkill() ;
+
+        MemberSkillCreateDTO memberSkillCreateDTO = new MemberSkillCreateDTO();
+        memberSkillCreateDTO.setMemberid(memberProfile.getUuid());
+        memberSkillCreateDTO.setSkillid(skill.getId());
+
+        final HttpRequest<MemberSkillCreateDTO> request = HttpRequest.POST("", memberSkillCreateDTO).basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+        final HttpResponse<MemberSkill> response = client.toBlocking().exchange(request, MemberSkill.class);
+
+        MemberSkill memberSkill = response.body();
+
+        assertEquals(memberSkill, response.body());
+        assertEquals(HttpStatus.CREATED, response.getStatus());
+        assertEquals(String.format("%s/%s", request.getPath(), memberSkill.getId()), response.getHeaders().get("location"));
+    }
+
+    @Test
+    void testCreateAnInvalidMemberSkill() {
+        MemberSkillCreateDTO memberSkillCreateDTO = new MemberSkillCreateDTO();
+
+        final HttpRequest<MemberSkillCreateDTO> request = HttpRequest.POST("", memberSkillCreateDTO).basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+        HttpClientResponseException responseException = assertThrows(HttpClientResponseException.class,
+                () -> client.toBlocking().exchange(request, Map.class));
+
+        JsonNode body = responseException.getResponse().getBody(JsonNode.class).orElse(null);
+        JsonNode errors = Objects.requireNonNull(body).get("_embedded").get("errors");
+        JsonNode href = Objects.requireNonNull(body).get("_links").get("self").get("href");
+        List<String> errorList = List.of(errors.get(0).get("message").asText(), errors.get(1).get("message").asText())
+                .stream().sorted().collect(Collectors.toList());
+        assertEquals("memberSkill.memberid: must not be null", errorList.get(0));
+        assertEquals("memberSkill.skillid: must not be null", errorList.get(1));
+        assertEquals(request.getPath(), href.asText());
+        assertEquals(HttpStatus.BAD_REQUEST, responseException.getStatus());
+    }
+
+    @Test
+    void testCreateANullMemberSkill() {
+        MemberSkill m = new MemberSkill(UUID.randomUUID(), UUID.randomUUID());
+
+        final HttpRequest<String> request = HttpRequest.POST("", "").basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+        HttpClientResponseException responseException = assertThrows(HttpClientResponseException.class,
+                () -> client.toBlocking().exchange(request, Map.class));
+
+        JsonNode body = responseException.getResponse().getBody(JsonNode.class).orElse(null);
+        JsonNode errors = Objects.requireNonNull(body).get("message");
+        JsonNode href = Objects.requireNonNull(body).get("_links").get("self").get("href");
+        assertEquals("Required Body [memberSkill] not specified", errors.asText());
+        assertEquals(request.getPath(), href.asText());
+        assertEquals(HttpStatus.BAD_REQUEST, responseException.getStatus());
+
+    }
+
+    @Test
+    void createAMemeberSkillForExistingMemberProfileO() {
+        MemberProfile memberProfile = createADefaultMemberProfile();
+        Skill skill = createADefaultSkill();
+
+        MemberSkillCreateDTO memberSkillCreateDTO = new MemberSkillCreateDTO();
+        memberSkillCreateDTO.setMemberid(UUID.randomUUID());
+        memberSkillCreateDTO.setSkillid(skill.getId());
+
+        final HttpRequest<MemberSkillCreateDTO> request = HttpRequest.POST("", memberSkillCreateDTO).basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+        HttpClientResponseException responseException = assertThrows(HttpClientResponseException.class,
+                () -> client.toBlocking().exchange(request, Map.class));
+
+        JsonNode body = responseException.getResponse().getBody(JsonNode.class).orElse(null);
+        String error = Objects.requireNonNull(body).get("message").asText();
+        String href = Objects.requireNonNull(body).get("_links").get("self").get("href").asText();
+
+        assertEquals(request.getPath(), href);
+        assertEquals(String.format("Member Profile %s doesn't exist",memberSkillCreateDTO.getMemberid()),error);
+
+    }
+
+    @Test
+    void createAMemeberSkillForExistingSkill() {
+        MemberProfile memberProfile = createADefaultMemberProfile();
+        Skill skill = createADefaultSkill();
+
+        MemberSkillCreateDTO memberSkillCreateDTO = new MemberSkillCreateDTO();
+        memberSkillCreateDTO.setMemberid(memberProfile.getUuid());
+        memberSkillCreateDTO.setSkillid(UUID.randomUUID());
+
+        final HttpRequest<MemberSkillCreateDTO> request = HttpRequest.POST("", memberSkillCreateDTO).basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+        HttpClientResponseException responseException = assertThrows(HttpClientResponseException.class,
+                () -> client.toBlocking().exchange(request, Map.class));
+
+        JsonNode body = responseException.getResponse().getBody(JsonNode.class).orElse(null);
+        String error = Objects.requireNonNull(body).get("message").asText();
+        String href = Objects.requireNonNull(body).get("_links").get("self").get("href").asText();
+
+        assertEquals(request.getPath(), href);
+        assertEquals(String.format("Skill %s doesn't exist",memberSkillCreateDTO.getSkillid()),error);
+
+    }
+
+    @Test
+    void createAMemeberSkillForExistingSkillAndMember() {
+        MemberProfile memberProfile = createADefaultMemberProfile();
+        Skill skill = createADefaultSkill();
+
+        MemberSkill memberSkill = createMemberSkill(memberProfile,skill);
+        MemberSkillCreateDTO memberSkillCreateDTO = new MemberSkillCreateDTO();
+        memberSkillCreateDTO.setMemberid(memberSkill.getMemberid());
+        memberSkillCreateDTO.setSkillid(memberSkill.getSkillid());
+
+        final HttpRequest<MemberSkillCreateDTO> request = HttpRequest.POST("", memberSkillCreateDTO).basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+        HttpClientResponseException responseException = assertThrows(HttpClientResponseException.class,
+                () -> client.toBlocking().exchange(request, Map.class));
+
+        JsonNode body = responseException.getResponse().getBody(JsonNode.class).orElse(null);
+        String error = Objects.requireNonNull(body).get("message").asText();
+        String href = Objects.requireNonNull(body).get("_links").get("self").get("href").asText();
+
+        assertEquals(request.getPath(), href);
+        assertEquals(String.format("Member %s already has this skill %s",memberSkillCreateDTO.getMemberid(),memberSkillCreateDTO.getSkillid()),error);
+
+    }
+
+
+    @Test
+    void deleteMemberSkillAsAdmin() {
+        MemberProfile memberProfile = createADefaultMemberProfile();
+        Skill skill = createADefaultSkill();
+
+        MemberSkill memberSkill = createMemberSkill(memberProfile,skill);
+
+        final HttpRequest<Object> request = HttpRequest.DELETE(memberSkill.getId().toString()).basicAuth(ADMIN_ROLE, ADMIN_ROLE);
+        final HttpResponse<String> response = client.toBlocking().exchange(request, String.class);
+
+        assertEquals(HttpStatus.OK, response.getStatus());
+    }
+
+    @Test
+    void deleteMemberSkillNotAsAdmin() {
+        MemberProfile memberProfile = createADefaultMemberProfile();
+        Skill skill = createADefaultSkill();
+
+        MemberSkill memberSkill = createMemberSkill(memberProfile,skill);
+
+        final HttpRequest<Object> request = HttpRequest.DELETE(memberSkill.getId().toString()).basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+
+        HttpClientResponseException responseException = assertThrows(HttpClientResponseException.class,
+                () -> client.toBlocking().exchange(request, String.class));
+
+        assertEquals(HttpStatus.FORBIDDEN, responseException.getStatus());
+
+    }
+
+    @Test
+    void testReadAllMemberSkills() {
+        MemberProfile memberProfile = createADefaultMemberProfile();
+        Skill skill = createADefaultSkill();
+
+        MemberSkill memberSkill = createMemberSkill(memberProfile,skill);
+
+        final HttpRequest<Object> request = HttpRequest.GET(String.format("/?memberid=%s&skillid=%s","","")).basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+        final HttpResponse<Set<MemberSkill>> response = client.toBlocking().exchange(request, Argument.setOf(MemberSkill.class));
+
+        assertEquals(Set.of(memberSkill), response.body());
+        assertEquals(HttpStatus.OK, response.getStatus());
+
+    }
+
+    @Test
+    void testReadMemberSkill() {
+        MemberProfile memberProfile = createADefaultMemberProfile();
+        Skill skill = createADefaultSkill();
+
+        MemberSkill memberSkill = createMemberSkill(memberProfile,skill);
+
+        final HttpRequest<Object> request = HttpRequest.GET(String.format("/%s", memberSkill.getId().toString())).basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+        final HttpResponse<MemberSkill> response = client.toBlocking().exchange(request, MemberSkill.class);
+
+        assertEquals(memberSkill, response.body());
+        assertEquals(HttpStatus.OK, response.getStatus());
+
+    }
+
+    @Test
+    void testReadMemberSkillNotFound() {
+
+        final HttpRequest<Object> request = HttpRequest.GET(String.format("/%s", UUID.randomUUID())).basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+        HttpClientResponseException responseException = assertThrows(HttpClientResponseException.class, () -> client.toBlocking().exchange(request, MemberSkill.class));
+
+        assertEquals(HttpStatus.NOT_FOUND, responseException.getStatus());
+    }
+
+    @Test
+    void testFindMemberSkills() {
+        MemberProfile memberProfile = createADefaultMemberProfile();
+        Skill skill = createADefaultSkill();
+
+        MemberSkill memberSkill = createMemberSkill(memberProfile,skill);
+
+        final HttpRequest<?> request = HttpRequest.GET(String.format("/?memberid=%s&skillid=%s", memberSkill.getMemberid(),
+                memberSkill.getSkillid())).basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+        final HttpResponse<Set<MemberSkill>> response = client.toBlocking().exchange(request, Argument.setOf(MemberSkill.class));
+
+        assertEquals(Set.of(memberSkill), response.body());
+        assertEquals(HttpStatus.OK, response.getStatus());
+
+    }
+
+    @Test
+    void testFindMemberSkillsByMemberId() {
+        MemberProfile memberProfile = createADefaultMemberProfile();
+        Skill skill = createADefaultSkill();
+
+        MemberSkill memberSkill = createMemberSkill(memberProfile,skill);
+
+        final HttpRequest<?> request = HttpRequest.GET(String.format("/?memberid=%s", memberSkill.getMemberid())).basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+        final HttpResponse<Set<MemberSkill>> response = client.toBlocking().exchange(request, Argument.setOf(MemberSkill.class));
+
+        assertEquals(Set.of(memberSkill), response.body());
+        assertEquals(HttpStatus.OK, response.getStatus());
+
+    }
+
+    @Test
+    void testFindMemberSkillsBySkillId() {
+        MemberProfile memberProfile = createADefaultMemberProfile();
+        Skill skill = createADefaultSkill();
+
+        MemberSkill memberSkill = createMemberSkill(memberProfile,skill);
+
+        final HttpRequest<?> request = HttpRequest.GET(String.format("/?skillid=%s", memberSkill.getSkillid())).basicAuth(MEMBER_ROLE, MEMBER_ROLE);
+        final HttpResponse<Set<MemberSkill>> response = client.toBlocking().exchange(request, Argument.setOf(MemberSkill.class));
+
+        assertEquals(Set.of(memberSkill), response.body());
+        assertEquals(HttpStatus.OK, response.getStatus());
+
+    }
+
+
+
+}
