+++ server/src/test/java/com/objectcomputing/checkins/services/member_skill/MemberSkillsCreateDTOTest.java
@@ -1,6 +1,5 @@
-package com.objectcomputing.checkins.services.memberskill;
+package com.objectcomputing.checkins.services.member_skill;
 
-import com.objectcomputing.checkins.services.memberSkill.MemberSkillCreateDTO;
 import io.micronaut.test.annotation.MicronautTest;
 import io.micronaut.validation.validator.Validator;
 import org.junit.jupiter.api.Test;
